requirements.txt
PyQt5
python-docx
pandas
